+++ gcdapi/backgroundservice.go
@@ -19,7 +19,7 @@ type BackgroundServiceBackgroundServiceEvent struct {
 	Timestamp                   float64                           `json:"timestamp"`                   // Timestamp of the event (in seconds).
 	Origin                      string                            `json:"origin"`                      // The origin this event belongs to.
 	ServiceWorkerRegistrationId string                            `json:"serviceWorkerRegistrationId"` // The Service Worker ID that initiated the event.
-	Service                     string                            `json:"service"`                     // The Background Service this event belongs to. enum values: backgroundFetch, backgroundSync
+	Service                     string                            `json:"service"`                     // The Background Service this event belongs to. enum values: backgroundFetch, backgroundSync, pushMessaging, notifications, paymentHandler
 	EventName                   string                            `json:"eventName"`                   // A description of the event.
 	InstanceId                  string                            `json:"instanceId"`                  // An identifier that groups related events together.
 	EventMetadata               []*BackgroundServiceEventMetadata `json:"eventMetadata"`               // A list of event-specific information.
@@ -30,7 +30,7 @@ type BackgroundServiceRecordingStateChangedEvent struct {
 	Method string `json:"method"`
 	Params struct {
 		IsRecording bool   `json:"isRecording"` //
-		Service     string `json:"service"`     //  enum values: backgroundFetch, backgroundSync
+		Service     string `json:"service"`     //  enum values: backgroundFetch, backgroundSync, pushMessaging, notifications, paymentHandler
 	} `json:"Params,omitempty"`
 }
 
@@ -52,7 +52,7 @@ func NewBackgroundService(target gcdmessage.ChromeTargeter) *BackgroundService {
 }
 
 type BackgroundServiceStartObservingParams struct {
-	//  enum values: backgroundFetch, backgroundSync
+	//  enum values: backgroundFetch, backgroundSync, pushMessaging, notifications, paymentHandler
 	Service string `json:"service"`
 }
 
@@ -62,7 +62,7 @@ func (c *BackgroundService) StartObservingWithParams(v *BackgroundServiceStartOb
 }
 
 // StartObserving - Enables event updates for the service.
-// service -  enum values: backgroundFetch, backgroundSync
+// service -  enum values: backgroundFetch, backgroundSync, pushMessaging, notifications, paymentHandler
 func (c *BackgroundService) StartObserving(service string) (*gcdmessage.ChromeResponse, error) {
 	var v BackgroundServiceStartObservingParams
 	v.Service = service
@@ -70,7 +70,7 @@ func (c *BackgroundService) StartObserving(service string) (*gcdmessage.ChromeRe
 }
 
 type BackgroundServiceStopObservingParams struct {
-	//  enum values: backgroundFetch, backgroundSync
+	//  enum values: backgroundFetch, backgroundSync, pushMessaging, notifications, paymentHandler
 	Service string `json:"service"`
 }
 
@@ -80,7 +80,7 @@ func (c *BackgroundService) StopObservingWithParams(v *BackgroundServiceStopObse
 }
 
 // StopObserving - Disables event updates for the service.
-// service -  enum values: backgroundFetch, backgroundSync
+// service -  enum values: backgroundFetch, backgroundSync, pushMessaging, notifications, paymentHandler
 func (c *BackgroundService) StopObserving(service string) (*gcdmessage.ChromeResponse, error) {
 	var v BackgroundServiceStopObservingParams
 	v.Service = service
@@ -90,7 +90,7 @@ func (c *BackgroundService) StopObserving(service string) (*gcdmessage.ChromeRes
 type BackgroundServiceSetRecordingParams struct {
 	//
 	ShouldRecord bool `json:"shouldRecord"`
-	//  enum values: backgroundFetch, backgroundSync
+	//  enum values: backgroundFetch, backgroundSync, pushMessaging, notifications, paymentHandler
 	Service string `json:"service"`
 }
 
@@ -101,7 +101,7 @@ func (c *BackgroundService) SetRecordingWithParams(v *BackgroundServiceSetRecord
 
 // SetRecording - Set the recording state for the service.
 // shouldRecord -
-// service -  enum values: backgroundFetch, backgroundSync
+// service -  enum values: backgroundFetch, backgroundSync, pushMessaging, notifications, paymentHandler
 func (c *BackgroundService) SetRecording(shouldRecord bool, service string) (*gcdmessage.ChromeResponse, error) {
 	var v BackgroundServiceSetRecordingParams
 	v.ShouldRecord = shouldRecord
@@ -110,7 +110,7 @@ func (c *BackgroundService) SetRecording(shouldRecord bool, service string) (*gc
 }
 
 type BackgroundServiceClearEventsParams struct {
-	//  enum values: backgroundFetch, backgroundSync
+	//  enum values: backgroundFetch, backgroundSync, pushMessaging, notifications, paymentHandler
 	Service string `json:"service"`
 }
 
@@ -120,7 +120,7 @@ func (c *BackgroundService) ClearEventsWithParams(v *BackgroundServiceClearEvent
 }
 
 // ClearEvents - Clears all stored data for the service.
-// service -  enum values: backgroundFetch, backgroundSync
+// service -  enum values: backgroundFetch, backgroundSync, pushMessaging, notifications, paymentHandler
 func (c *BackgroundService) ClearEvents(service string) (*gcdmessage.ChromeResponse, error) {
 	var v BackgroundServiceClearEventsParams
 	v.Service = service