+++ gcdapi/webaudio.go
@@ -9,10 +9,12 @@ import (
 	"github.com/wirepair/gcd/gcdmessage"
 )
 
-// Fields in AudioContext that change in real-time. These are not updated on OfflineAudioContext.
+// Fields in AudioContext that change in real-time.
 type WebAudioContextRealtimeData struct {
-	CurrentTime    float64 `json:"currentTime,omitempty"`    // The current context time in second in BaseAudioContext.
-	RenderCapacity float64 `json:"renderCapacity,omitempty"` // The time spent on rendering graph divided by render qunatum duration, and multiplied by 100. 100 means the audio renderer reached the full capacity and glitch may occur.
+	CurrentTime              float64 `json:"currentTime"`              // The current context time in second in BaseAudioContext.
+	RenderCapacity           float64 `json:"renderCapacity"`           // The time spent on rendering graph divided by render qunatum duration, and multiplied by 100. 100 means the audio renderer reached the full capacity and glitch may occur.
+	CallbackIntervalMean     float64 `json:"callbackIntervalMean"`     // A running mean of callback interval.
+	CallbackIntervalVariance float64 `json:"callbackIntervalVariance"` // A running variance of callback interval.
 }
 
 // Protocol object for BaseAudioContext